Done with tests for now.
There still need to be more tests written for `Game.SlowChess.Move`
and the special moves.

A lot of the movement needs more tests that validate the absence of
illegal behaviour. The FIDE tests do a good job of making sure at
least the legal versions of moves are being captured.

I'm bullshitting. I just want to play with game trees.